adding more files

+++ pages/events/[slug].js
@@ -1,20 +1,35 @@
+import {ToastContainer, toast} from 'react-toastify';
+import {useRouter} from 'next/router';
 import { FaPencilAlt, FaTimes } from "react-icons/fa";
 import Layout from "@/components/Layout";
 import Link from "next/link";
 import Image from "next/image";
 import { API_URL } from "@/config/index";
+
 import styles from "@/styles/Event.module.css";
+import 'react-toastify/dist/ReactToastify.css';
 
 export default function EventPage({ evt }) {
-  const deleteEvent = async id => {
-    const res = await fetch(`${API_URL}/api/events/${id}`, {
-      method: "DELETE",
-    });
-    const data = await res.json();
-    console.log(data);
+  const { attributes } = evt;
+  const router = useRouter();
+
+  const deleteEvent = async () => {
+    if(confirm("Are you sure you want to delete this event?")) {
+      const res = await fetch(`${API_URL}/api/events/${evt.id}`, {
+        method: "DELETE",
+      });
+
+      const json = await res.json();
+      const event = json.data;
+
+      if (!res.ok) {
+        toast.error(data.message);
+      } else {
+        router.push(`/events`);
+      }
+    }
   };
   
-  const { attributes } = evt;
   return (
     <Layout>
       <div className={styles.event}>
@@ -30,17 +45,18 @@ export default function EventPage({ evt }) {
           {new Date(attributes.date).toLocaleDateString('en-GB')} at {attributes.time}
         </span>
         <h1>{attributes.name}</h1>
+        <ToastContainer/>
         {attributes.image && (
           <div className={styles.image}>
             <Image
-              src={attributes.image.data.attributes.formats.large.url}
+              src={attributes.image.data ? attributes.image.data.attributes.formats.large.url : "/images/event-default.png"}
               width={960}
               height={600}
               alt="event image"
             ></Image>
           </div>
         )}
-        <h3>Performerse:</h3>
+        <h3>Performers:</h3>
         <p>{attributes.performers}</p>
         <h3>Description</h3>
         <p>{attributes.description}</p>
